minor bugfix

+++ src/minimetro/WorldPanel.java
@@ -151,7 +151,7 @@ public class WorldPanel extends JPanel implements MouseListener,
                 prevRow = Integer.MAX_VALUE;
             }
             case STATION -> {
-                world.toggleStation(currentRow, currentCol);
+                world.toggleStation((int) currentRow, (int) currentCol);
                 repaint();
             }
             }
